Rewritten the computation utils to a. remove x0cr and rx0, so that EpileptorDP2D has the same z function as the rest of the models. b. to substitute the special EpileptorDP2D b parameter with d-b c. get (default) parameter values from constants.py
Adjusted model_configuration and its service to:
a. get all these parameters,
b. to work with x0 instead of x0_values,
c. allow for x1eq beyond the separatrix

Adjusted lsa_service to check for the x1eq equilibria

Extended the nullclines plot up to x1=+2.0

Adjusted all the rest of the module to work with the new set of model
parameters

Tests run OK